// src/component-library/Map/sampleData/markers.d.ts
export declare const BOULAY_ATTACK: {
    geohash: string;
    type: string;
    uri: string;
    name: string;
};
export declare const ARGA_ATTACK: {
    geohash: string;
    type: string;
    uri: string;
    name: string;
};
export declare const MADE_UP_MARKER: {
    geohash: string;
    type: string;
    uri: string;
    name: string;
};